+++ kleinman.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: kleinman.c,v 1.3 2002/01/08 23:17:36 oysteijo Exp $
+ * $Id: kleinman.c,v 1.4 2006/09/21 22:24:05 Superfly_Jon Exp $
  */
 
 #if HAVE_CONFIG_H
@@ -50,7 +50,7 @@ KleinmanCount (int nPipOnRoll, int nPipNotOnRoll)
 
   rK = (double) (nDiff + 4) / (2 * sqrt( nSum - 4 ));
 
-  return 0.5 * (1.0 + erf( rK ));
+  return 0.5f * (1.0f + (float)erf( rK ));
 }
 
 #ifdef STANDALONE
